+++ src/common/av1.cpp
@@ -43,7 +43,7 @@ class parser_private_c {
   frame_t current_frame;
   std::deque<frame_t> frames;
   uint64_t frame_number{}, num_units_in_display_tick{}, time_scale{}, num_ticks_per_picture{};
-  boost::rational<uint64_t> forced_default_duration, bitstream_default_duration{1'000'000'000ull, 25};
+  mtx_mp_rational_t forced_default_duration, bitstream_default_duration{1'000'000'000ull, 25};
 
   memory_cptr sequence_header_obu;
   std::vector<memory_cptr> metadata_obus;
@@ -79,7 +79,7 @@ parser_c::get_obu_type_name(unsigned int obu_type) {
 }
 
 void
-parser_c::set_default_duration(boost::rational<uint64_t> default_duration) {
+parser_c::set_default_duration(mtx_mp_rational_t default_duration) {
   p->forced_default_duration = default_duration;
 }
 
@@ -90,9 +90,9 @@ parser_c::set_parse_sequence_header_obus_only(bool parse_sequence_header_obus_on
 
 uint64_t
 parser_c::get_next_timestamp() {
-  auto &duration = p->forced_default_duration ? p->forced_default_duration : p->bitstream_default_duration;
+  auto duration = p->forced_default_duration ? p->forced_default_duration : p->bitstream_default_duration;
 
-  return boost::rational_cast<int64_t>(duration * p->frame_number++);
+  return static_cast<int64_t>(duration * p->frame_number++);
 }
 
 uint64_t
@@ -242,7 +242,7 @@ parser_c::parse_timing_info(mtx::bits::reader_c &r) {
                          p->time_scale,
                          p->equal_picture_interval,
                          p->num_ticks_per_picture,
-                         mtx::string::format_timestamp(boost::rational_cast<uint64_t>(p->bitstream_default_duration))));
+                         mtx::string::format_timestamp(static_cast<uint64_t>(p->bitstream_default_duration))));
 }
 
 void