+++ src/common/common.h
@@ -45,7 +45,6 @@
 
 #include <boost/algorithm/string.hpp>
 #include <boost/multiprecision/gmp.hpp>
-#include <boost/rational.hpp>
 
 namespace balg = boost::algorithm;