+++ src/common/math.cpp
@@ -51,8 +51,8 @@ int_to_double(int64_t value) {
 mtx_mp_rational_t
 clamp_values_to(mtx_mp_rational_t const &r,
                 int64_t max_value) {
-  auto num = r.numerator();
-  auto den = r.denominator();
+  auto num = boost::multiprecision::numerator(r);
+  auto den = boost::multiprecision::denominator(r);
 
   if (!num || !den || ((num <= max_value) && (den <= max_value)))
     return r;
@@ -60,8 +60,8 @@ clamp_values_to(mtx_mp_rational_t const &r,
   // 333 / 1000 , clamp = 500, mul = 1/2 = 1/(max/clamp) = clamp/max
 
   auto mult = mtx_mp_rational_t{ max_value, std::max(num, den) };
-  den       = boost::rational_cast<int64_t>(den * mult);
+  den       = static_cast<int64_t>(den * mult);
 
-  return { boost::rational_cast<int64_t>(num * mult), den ? den : 1 };
+  return { static_cast<int64_t>(num * mult), den ? den : 1 };
 }
 }