+++ src/common/math_fwd.h
@@ -19,5 +19,5 @@ namespace mtx::math {
 
 }
 
-using mtx_mp_rational_t = boost::rational<int64_t>;
-using mtx_mp_int_t      = boost::multiprecision::number<boost::multiprecision::gmp_int, boost::multiprecision::et_off>;
+using mtx_mp_rational_t = boost::multiprecision::number<boost::multiprecision::gmp_rational, boost::multiprecision::et_off>;
+using mtx_mp_int_t      = boost::multiprecision::number<boost::multiprecision::gmp_int,      boost::multiprecision::et_off>;